Fix bugs: CommandWords toString does not enforce strict ordering. SesseionData freeResources nullPointer.

+++ src/main/java/seedu/carvicim/storage/session/SessionData.java
@@ -183,7 +183,9 @@ public class SessionData {
     public void freeResources() {
         workbook = null;
         importFile = null;
-        tempFile.delete();
+        if (importFile != null) {
+            tempFile.delete();
+        }
         tempFile = null;
         saveFile = null;
         unreviewedJobEntries.clear();
